c313: extract BenchmarksSummary tab cluster (ProtocolBox + Stat + SceneForest + PairedTable + MultiAxisBatterySection + METHOD_LABEL/COLOR) to own file (#441 P0 2.2)
Benchmarks.tsx 3153 LOC -> 2645 LOC (-508 LOC). Benchmarks chunk 72.11 KB -> 61.92 KB (-10.2 KB).
New chunk: BenchmarksSummary-COLlfKGG.js 11.13 KB.
Lifted LABELLED_SCENES const to pages/benchmarks/shared.ts.
Removed unused MethodStatistics + SceneMethodStats type imports from Benchmarks.tsx.
typecheck clean, 44/44 tests pass.
