+++ frontend/src/pages/benchmarks/shared.ts
@@ -1,0 +1,8 @@
+export const LABELLED_SCENES = [
+  "indian-pines-corrected",
+  "salinas-corrected",
+  "salinas-a-corrected",
+  "pavia-university",
+  "kennedy-space-center",
+  "botswana",
+];
